// web/src/shared/api/generated/model/subjectResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface SubjectResult {
  id: string;
  name: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  order: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  unitCount: number | string;
}